// src/components/quotation/types.ts
export interface QuotationItem {
    description: string;
    quantity: string;
    unitPrice: string;
}

export interface FormState {
    toAddress: string;
    name: string;
    refNo: string;
    mobile: string;
    date: string;
    brand: string;
    model: string;
    machineType: string;
    vat: string;
}